# chore: updated ESLint to v9 with flat config