--- a/apps/site/package.json
+++ b/apps/site/package.json
@@ -23,7 +23,7 @@
     "start": "cross-env NODE_NO_WARNINGS=1 next start",
     "deploy": "cross-env NEXT_PUBLIC_STATIC_EXPORT=true npm run build",
     "check-types": "tsc --noEmit",
-    "lint:js": "eslint \"**/*.{js,mjs,ts,tsx}\" --cache --cache-strategy=content --cache-location=.eslintjscache",
+    "lint:js": "eslint \"**/*.{js,mjs,ts,tsx}\"",
     "lint:md": "eslint \"**/*.md?(x)\" --cache --cache-strategy=content --cache-location=.eslintmdcache",
     "lint:css": "stylelint \"**/*.css\" --allow-empty-input --cache --cache-strategy=content --cache-location=.stylelintcache",
     "lint": "turbo run lint:md lint:js lint:css",
@@ -89,6 +89,8 @@
     "vfile-matter": "~5.0.0"
   },
   "devDependencies": {
+    "@eslint/compat": "^1.1.1",
+    "@next/eslint-plugin-next": "^14.2.8",
     "@storybook/addon-controls": "~8.2.7",
     "@storybook/addon-interactions": "~8.2.7",
     "@storybook/addon-themes": "~8.2.7",
@@ -101,14 +103,16 @@
     "@types/react": "^18.3.5",
     "@types/react-dom": "^18.3.0",
     "@types/semver": "~7.5.8",
-    "@typescript-eslint/eslint-plugin": "7.11.0",
-    "@typescript-eslint/parser": "7.11.0",
-    "eslint": "8.57.0",
+    "eslint": "^9.9.1",
     "eslint-config-next": "~14.2.3",
     "eslint-config-prettier": "9.1.0",
+    "eslint-import-resolver-typescript": "^3.6.3",
+    "eslint-plugin-import-x": "^4.2.1",
     "eslint-plugin-mdx": "3.1.5",
     "eslint-plugin-no-relative-import-paths": "^1.5.3",
-    "eslint-plugin-storybook": "0.8.0",
+    "eslint-plugin-react": "^7.35.2",
+    "eslint-plugin-react-hooks": "5.1.0-rc-4c58fce7-20240904",
+    "eslint-plugin-storybook": "0.9.0--canary.156.da7873a.0",
     "handlebars": "4.7.8",
     "jest": "29.7.0",
     "jest-environment-jsdom": "29.7.0",
@@ -120,6 +124,10 @@
     "stylelint-config-standard": "36.0.0",
     "stylelint-order": "6.0.4",
     "stylelint-selector-bem-pattern": "4.0.0",
+    "typescript-eslint": "^8.4.0",
     "user-agent-data-types": "0.4.2"
+  },
+  "overrides": {
+    "eslint": "$eslint"
   }
 }